moved abi encoder constants into utils dir
