+++ packages/utils/src/abi_encoder/abstract_data_types/dependent_data_type.ts
@@ -3,7 +3,7 @@ import * as ethUtil from 'ethereumjs-util';
 import * as _ from 'lodash';
 
 import { CalldataBlock, DependentCalldataBlock, RawCalldata } from '../calldata';
-import * as Constants from '../constants';
+import * as Constants from '../utils/constants';
 import { DecodingRules } from '../utils/rules';
 
 import { DataType } from './data_type';
