+++ packages/utils/src/abi_encoder/calldata/dependent_calldata_block.ts
@@ -1,6 +1,6 @@
 import * as ethUtil from 'ethereumjs-util';
 
-import * as Constants from '../constants';
+import * as Constants from '../utils/constants';
 
 import { CalldataBlock } from './calldata_block';
