+++ packages/utils/src/abi_encoder/calldata/raw_calldata.ts
@@ -1,6 +1,6 @@
 import * as ethUtil from 'ethereumjs-util';
 
-import * as Constants from '../constants';
+import * as Constants from '../utils/constants';
 import { Queue } from '../utils/queue';
 
 export class RawCalldata {
