+++ packages/utils/src/abi_encoder/evm_data_types/array.ts
@@ -1,7 +1,7 @@
 import { DataItem } from 'ethereum-types';
 
 import { DataTypeFactory, MemberDataType } from '../abstract_data_types';
-import * as Constants from '../constants';
+import * as Constants from '../utils/constants';
 
 export class Array extends MemberDataType {
     private static readonly _matcher = RegExp('^(.+)\\[([0-9]*)\\]$');
