+++ packages/utils/src/abi_encoder/evm_data_types/dynamic_bytes.ts
@@ -5,7 +5,7 @@ import * as _ from 'lodash';
 
 import { DataTypeFactory, PayloadDataType } from '../abstract_data_types';
 import { RawCalldata } from '../calldata';
-import * as Constants from '../constants';
+import * as Constants from '../utils/constants';
 
 export class DynamicBytes extends PayloadDataType {
     private static readonly _SIZE_KNOWN_AT_COMPILE_TIME: boolean = false;
