+++ packages/utils/src/abi_encoder/evm_data_types/number.ts
@@ -5,7 +5,7 @@ import * as _ from 'lodash';
 import { BigNumber } from '../../configured_bignumber';
 import { DataTypeFactory, PayloadDataType } from '../abstract_data_types';
 import { RawCalldata } from '../calldata';
-import * as Constants from '../constants';
+import * as Constants from '../utils/constants';
 
 export abstract class Number extends PayloadDataType {
     private static readonly _SIZE_KNOWN_AT_COMPILE_TIME: boolean = true;
