+++ packages/utils/src/abi_encoder/evm_data_types/string.ts
@@ -5,7 +5,7 @@ import * as _ from 'lodash';
 
 import { DataTypeFactory, PayloadDataType } from '../abstract_data_types';
 import { RawCalldata } from '../calldata';
-import * as Constants from '../constants';
+import * as Constants from '../utils/constants';
 
 export class String extends PayloadDataType {
     private static readonly _SIZE_KNOWN_AT_COMPILE_TIME: boolean = false;
